requirements.txt
affine==2.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
bleach==6.3.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cligj==0.7.2
contourpy==1.3.3
cryptography==46.0.5
cuda-bindings==13.2.0
cuda-pathfinder==1.4.4
cuda-toolkit==13.0.2
cycler==0.12.1
earthengine-api==1.7.18
fastapi==0.135.2
filelock==3.25.2
fonttools==4.62.1
fsspec==2026.2.0
google-api-core==2.30.0
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-cloud-core==2.5.0
google-cloud-storage==3.10.1
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.73.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface_hub==1.8.0
idna==3.11
Jinja2==3.1.6
kaggle==2.0.0
kagglesdk==0.1.16
kiwisolver==1.5.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.3
nvidia-cublas==13.1.0.3
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-runtime==13.0.96
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.4.66
nvidia-cusparse==12.6.3.3
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
opencv-python==4.13.0.92
packaging==26.0
pillow==12.1.1
proto-plus==1.27.1
protobuf==6.33.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.3
rasterio==1.5.0
requests==2.32.5
rich==14.3.3
safetensors==0.7.0
segmentation_models_pytorch==0.5.0
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
starlette==1.0.0
sympy==1.14.0
text-unidecode==1.3
timm==1.0.26
torch==2.11.0
torchvision==0.26.0
tqdm==4.67.3
triton==3.6.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.42.0
webencodings==0.5.1
